requirements.txt
django==1.10
python-decouple==3.1
Pillow==5.3.0
djangorestframework==3.8.2
markdown==3.0.1
django-filter==2.0.0